@@ -2,14 +2,14 @@ name: android_intent_example
 description: Demonstrates how to use the android_intent plugin.
 
 environment:
-  sdk: ">=2.12.0 <3.0.0"
+  sdk: ">=2.18.0 <4.0.0"
+  flutter: ">=3.3.0"
 
 dependencies:
   flutter:
     sdk: flutter
   platform: ^3.1.0
-  android_intent_plus:
-    path: ../
+  android_intent_plus: ^4.0.0
 
 dev_dependencies:
   flutter_driver:
