@@ -15,15 +15,18 @@ void main() {
     setUp(() async {
       methodChannelDeviceInfo = MethodChannelDeviceInfo();
 
-      methodChannelDeviceInfo.channel
-          .setMockMethodCallHandler((MethodCall methodCall) async {
-        switch (methodCall.method) {
-          case 'getDeviceInfo':
-            return {'device_info': 'is_fake'};
-          default:
-            return null;
-        }
-      });
+      TestDefaultBinaryMessengerBinding.instance.defaultBinaryMessenger
+          .setMockMethodCallHandler(
+        methodChannelDeviceInfo.channel,
+        (MethodCall methodCall) async {
+          switch (methodCall.method) {
+            case 'getDeviceInfo':
+              return {'device_info': 'is_fake'};
+            default:
+              return null;
+          }
+        },
+      );
     });
 
     test('deviceInfo', () async {
