@@ -17,10 +17,10 @@ void main() {
     final info = await PackageInfo.fromPlatform();
     // These tests are based on the example app. The tests should be updated if any related info changes.
     if (kIsWeb) {
-      expect(info.appName, 'package_info_example');
+      expect(info.appName, 'package_info_plus_example');
       expect(info.buildNumber, '4');
       expect(info.buildSignature, isEmpty);
-      expect(info.packageName, 'package_info_example');
+      expect(info.packageName, 'package_info_plus_example');
       expect(info.version, '1.2.3');
       expect(info.installerStore, null);
     } else {
@@ -68,7 +68,7 @@ void main() {
     await tester.pumpWidget(const MyApp());
     await tester.pumpAndSettle();
     if (kIsWeb) {
-      expect(find.text('package_info_example'), findsNWidgets(2));
+      expect(find.text('package_info_plus_example'), findsNWidgets(2));
       expect(find.text('1.2.3'), findsOneWidget);
       expect(find.text('4'), findsOneWidget);
       expect(find.text('Not set'), findsOneWidget);
