@@ -6,106 +6,108 @@
 // - github.com/chromium/chromium/src/base/file_version_info_win.cc
 // - github.com/timsneath/win32/example/filever.dart
 
-part of package_info_plus_windows;
+import 'dart:ffi';
+import 'dart:io';
 
-class _LANGANDCODEPAGE extends Struct {
-  @Uint16()
+import 'package:ffi/ffi.dart';
+import 'package:win32/win32.dart';
+
+class LANGANDCODEPAGE extends Struct {
+  @WORD()
   external int wLanguage;
 
-  @Uint16()
+  @WORD()
   external int wCodePage;
 }
 
-final _kernel32 = DynamicLibrary.open('kernel32.dll');
-// ignore: non_constant_identifier_names
-final _GetUserDefaultLangID = _kernel32
-    .lookupFunction<Uint16 Function(), int Function()>('GetUserDefaultLangID');
-
-class _FileVersionInfoData {
-  _FileVersionInfoData({this.lpBlock, this.lpLang});
-  final Pointer<Uint8>? lpBlock;
-  final Pointer<_LANGANDCODEPAGE>? lpLang;
+class FileVersionInfoData {
+  const FileVersionInfoData({required this.lpBlock, required this.lpLang});
+  final Pointer<BYTE> lpBlock;
+  final Pointer<LANGANDCODEPAGE> lpLang;
 }
 
-class _FileVersionInfo {
+class FileVersionInfo {
   final String filePath;
-  final _FileVersionInfoData _data;
+  final FileVersionInfoData _data;
 
-  _FileVersionInfo(this.filePath) : _data = _getData(filePath);
+  FileVersionInfo(this.filePath) : _data = getData(filePath);
 
-  void dispose() => calloc.free(_data.lpBlock!);
+  void dispose() => free(_data.lpBlock);
 
-  String? get companyName => getValue('CompanyName');
-  String? get companyShortName => getValue('CompanyShortName');
-  String? get productName => getValue('ProductName');
-  String? get productShortName => getValue('ProductShortName');
-  String? get internalName => getValue('InternalName');
-  String? get productVersion => getValue('ProductVersion');
-  String? get specialBuild => getValue('SpecialBuild');
-  String? get originalFilename => getValue('OriginalFilename');
-  String? get fileDescription => getValue('FileDescription');
-  String? get fileVersion => getValue('FileVersion');
+  String get companyName => getValue('CompanyName');
+  String get companyShortName => getValue('CompanyShortName');
+  String get productName => getValue('ProductName');
+  String get productShortName => getValue('ProductShortName');
+  String get internalName => getValue('InternalName');
+  String get productVersion => getValue('ProductVersion');
+  String get specialBuild => getValue('SpecialBuild');
+  String get originalFilename => getValue('OriginalFilename');
+  String get fileDescription => getValue('FileDescription');
+  String get fileVersion => getValue('FileVersion');
 
-  String? getValue(String name) {
+  String getValue(String name) {
     final langCodepages = [
       // try the language and codepage from the EXE
-      [_data.lpLang!.ref.wLanguage, _data.lpLang!.ref.wCodePage],
+      [_data.lpLang.ref.wLanguage, _data.lpLang.ref.wCodePage],
       // try the default language and codepage from the EXE
-      [_GetUserDefaultLangID(), _data.lpLang!.ref.wCodePage],
+      [GetUserDefaultLangID(), _data.lpLang.ref.wCodePage],
       // try the language from the EXE and Latin codepage (most common)
-      [_data.lpLang!.ref.wLanguage, 1252],
+      [_data.lpLang.ref.wLanguage, 1252],
       // try the default language and Latin codepage (most common)
-      [_GetUserDefaultLangID(), 1252],
+      [GetUserDefaultLangID(), 1252],
     ];
 
-    String? value;
-    final Pointer<IntPtr> lplpBuffer = calloc<IntPtr>();
-    final Pointer<Uint32> puLen = calloc<Uint32>();
-
+    final lplpBuffer = calloc<LPWSTR>();
+    final puLen = calloc<UINT>();
     String toHex4(int val) => val.toRadixString(16).padLeft(4, '0');
 
-    for (final langCodepage in langCodepages) {
-      final lang = toHex4(langCodepage[0]);
-      final codepage = toHex4(langCodepage[1]);
-      final lpSubBlock = TEXT('\\StringFileInfo\\$lang$codepage\\$name');
-      final res =
-          VerQueryValue(_data.lpBlock!, lpSubBlock, lplpBuffer.cast(), puLen);
-      calloc.free(lpSubBlock);
+    try {
+      for (final langCodepage in langCodepages) {
+        final lang = toHex4(langCodepage[0]);
+        final codepage = toHex4(langCodepage[1]);
+        final lpSubBlock = TEXT('\\StringFileInfo\\$lang$codepage\\$name');
+        final res =
+            VerQueryValue(_data.lpBlock, lpSubBlock, lplpBuffer.cast(), puLen);
+        free(lpSubBlock);
 
-      if (res != 0 && lplpBuffer.value != 0 && puLen.value > 0) {
-        final ptr = Pointer<Utf16>.fromAddress(lplpBuffer.value);
-        value = ptr.toDartString();
-        break;
+        if (res != 0 && lplpBuffer.address != 0 && puLen.value > 0) {
+          return lplpBuffer.value.toDartString();
+        }
       }
-    }
 
-    calloc.free(lplpBuffer);
-    calloc.free(puLen);
-    return value;
+      return '';
+    } finally {
+      free(lplpBuffer);
+      free(puLen);
+    }
   }
 
-  static _FileVersionInfoData _getData(String filePath) {
+  static FileVersionInfoData getData(String filePath) {
+    if (!File(filePath).existsSync()) {
+      throw ArgumentError.value(filePath, 'filePath', 'File not present');
+    }
+
     final lptstrFilename = TEXT(filePath);
-    final lpdwDummy = calloc<Uint32>();
-    final dwBlockSize = GetFileVersionInfoSize(lptstrFilename, lpdwDummy);
-    final lpBlock = calloc<Uint8>(dwBlockSize);
-    if (GetFileVersionInfo(lptstrFilename, 0, dwBlockSize, lpBlock) == 0) {
-      throw WindowsException(HRESULT_FROM_WIN32(GetLastError()));
-    }
+    final dwLen = GetFileVersionInfoSize(lptstrFilename, nullptr);
+
+    final lpData = calloc<BYTE>(dwLen); // freed by the dispose() method
     final lpSubBlock = TEXT(r'\VarFileInfo\Translation');
-    final lpTranslate = calloc<IntPtr>();
-    if (VerQueryValue(lpBlock, lpSubBlock, lpTranslate.cast(), lpdwDummy) ==
-        0) {
-      throw WindowsException(HRESULT_FROM_WIN32(GetLastError()));
+    final lpTranslate = calloc<Pointer<LANGANDCODEPAGE>>();
+    final puLen = calloc<UINT>();
+    try {
+      if (GetFileVersionInfo(lptstrFilename, NULL, dwLen, lpData) == 0) {
+        throw WindowsException(HRESULT_FROM_WIN32(GetLastError()));
+      }
+
+      if (VerQueryValue(lpData, lpSubBlock, lpTranslate.cast(), puLen) == 0) {
+        throw WindowsException(HRESULT_FROM_WIN32(GetLastError()));
+      }
+      return FileVersionInfoData(lpBlock: lpData, lpLang: lpTranslate.value);
+    } finally {
+      free(lptstrFilename);
+      free(lpTranslate);
+      free(lpSubBlock);
+      free(puLen);
     }
-    final data = _FileVersionInfoData(
-      lpBlock: lpBlock,
-      lpLang: Pointer<_LANGANDCODEPAGE>.fromAddress(lpTranslate.value),
-    );
-    calloc.free(lptstrFilename);
-    calloc.free(lpTranslate);
-    calloc.free(lpSubBlock);
-    calloc.free(lpdwDummy);
-    return data;
   }
 }
