@@ -23,7 +23,9 @@ class PackageInfoPlusWebPlugin extends PackageInfoPlatform {
   /// Get version.json full url.
   Uri versionJsonUrl(String baseUrl, int cacheBuster) {
     final baseUri = Uri.parse(baseUrl);
-    final originPath = '${baseUri._origin}${baseUri.path}';
+    final regExp = RegExp(r'\w+\.html.*');
+    final originPath =
+        '${baseUri._origin}${baseUri.path.replaceAll(regExp, '')}';
     final versionJson = 'version.json?cachebuster=$cacheBuster';
     return Uri.parse(originPath.endsWith('/')
         ? '$originPath$versionJson'
