@@ -1,19 +1,16 @@
 /// The Windows implementation of `package_info_plus`.
 library package_info_plus_windows;
 
-import 'dart:ffi';
 import 'dart:io';
 
-import 'package:ffi/ffi.dart';
 import 'package:package_info_plus_platform_interface/package_info_data.dart';
 import 'package:package_info_plus_platform_interface/package_info_platform_interface.dart';
-import 'package:win32/win32.dart';
 
-part 'file_version_info.dart';
+import 'file_version_info.dart';
 
 /// The Windows implementation of [PackageInfoPlatform].
 class PackageInfoPlusWindowsPlugin extends PackageInfoPlatform {
-  /// Register this dart class as the platform implementation for linux
+  /// Register this dart class as the platform implementation for Windows
   static void registerWith() {
     PackageInfoPlatform.instance = PackageInfoPlusWindowsPlugin();
   }
@@ -22,11 +19,18 @@ class PackageInfoPlusWindowsPlugin extends PackageInfoPlatform {
   /// appName, packageName, version, buildNumber
   @override
   Future<PackageInfoData> getAll() {
-    final info = _FileVersionInfo(Platform.resolvedExecutable);
-    final versions = info.productVersion!.split('+');
+    String resolvedExecutable = Platform.resolvedExecutable;
+
+    /// Workaround for https://github.com/dart-lang/sdk/issues/52309
+    if (resolvedExecutable.startsWith(r"UNC\")) {
+      resolvedExecutable = resolvedExecutable.replaceFirst(r"UNC\", r"\\");
+    }
+
+    final info = FileVersionInfo(Platform.resolvedExecutable);
+    final versions = info.productVersion.split('+');
     final data = PackageInfoData(
-      appName: info.productName ?? '',
-      packageName: info.internalName ?? '',
+      appName: info.productName,
+      packageName: info.internalName,
       version: versions.getOrNull(0) ?? '',
       buildNumber: versions.getOrNull(1) ?? '',
       buildSignature: '',
@@ -38,5 +42,6 @@ class PackageInfoPlusWindowsPlugin extends PackageInfoPlatform {
 
 extension _GetOrNull<T> on List<T> {
   T? getOrNull(int index) => _checkIndex(index) ? this[index] : null;
+
   bool _checkIndex(int index) => index >= 0 && index < length;
 }
