@@ -1,9 +1,6 @@
 // Copyright 2019, the Chromium project authors.  Please see the AUTHORS file
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
-
-// @dart=2.9
-
 import 'dart:async';
 import 'dart:io';
 import 'package:flutter_test/flutter_test.dart';
@@ -14,10 +11,10 @@ void main() {
   IntegrationTestWidgetsFlutterBinding.ensureInitialized();
 
   // NOTE: The accelerometer events are not returned on iOS simulators.
-  testWidgets('Can subscript to accelerometerEvents and get non-null events',
+  testWidgets('Can subscribe to accelerometerEvents and get non-null events',
       (WidgetTester tester) async {
     final completer = Completer<AccelerometerEvent>();
-    StreamSubscription<AccelerometerEvent> subscription;
+    late StreamSubscription<AccelerometerEvent> subscription;
     subscription = accelerometerEvents.listen((AccelerometerEvent event) {
       completer.complete(event);
       subscription.cancel();
