@@ -54,21 +54,21 @@ class _MyHomePageState extends State<MyHomePage> {
   static const int _snakeColumns = 20;
   static const double _snakeCellSize = 10.0;
 
-  List<double>? _accelerometerValues;
   List<double>? _userAccelerometerValues;
+  List<double>? _accelerometerValues;
   List<double>? _gyroscopeValues;
   List<double>? _magnetometerValues;
   final _streamSubscriptions = <StreamSubscription<dynamic>>[];
 
   @override
   Widget build(BuildContext context) {
+    final userAccelerometer = _userAccelerometerValues
+        ?.map((double v) => v.toStringAsFixed(1))
+        .toList();
     final accelerometer =
         _accelerometerValues?.map((double v) => v.toStringAsFixed(1)).toList();
     final gyroscope =
         _gyroscopeValues?.map((double v) => v.toStringAsFixed(1)).toList();
-    final userAccelerometer = _userAccelerometerValues
-        ?.map((double v) => v.toStringAsFixed(1))
-        .toList();
     final magnetometer =
         _magnetometerValues?.map((double v) => v.toStringAsFixed(1)).toList();
 
@@ -101,7 +101,7 @@ class _MyHomePageState extends State<MyHomePage> {
             child: Row(
               mainAxisAlignment: MainAxisAlignment.spaceBetween,
               children: <Widget>[
-                Text('Accelerometer: $accelerometer'),
+                Text('UserAccelerometer: $userAccelerometer'),
               ],
             ),
           ),
@@ -110,7 +110,7 @@ class _MyHomePageState extends State<MyHomePage> {
             child: Row(
               mainAxisAlignment: MainAxisAlignment.spaceBetween,
               children: <Widget>[
-                Text('UserAccelerometer: $userAccelerometer'),
+                Text('Accelerometer: $accelerometer'),
               ],
             ),
           ),
@@ -149,10 +149,10 @@ class _MyHomePageState extends State<MyHomePage> {
   void initState() {
     super.initState();
     _streamSubscriptions.add(
-      accelerometerEvents.listen(
-        (AccelerometerEvent event) {
+      userAccelerometerEvents.listen(
+        (UserAccelerometerEvent event) {
           setState(() {
-            _accelerometerValues = <double>[event.x, event.y, event.z];
+            _userAccelerometerValues = <double>[event.x, event.y, event.z];
           });
         },
         onError: (e) {
@@ -170,10 +170,10 @@ class _MyHomePageState extends State<MyHomePage> {
       ),
     );
     _streamSubscriptions.add(
-      gyroscopeEvents.listen(
-        (GyroscopeEvent event) {
+      accelerometerEvents.listen(
+        (AccelerometerEvent event) {
           setState(() {
-            _gyroscopeValues = <double>[event.x, event.y, event.z];
+            _accelerometerValues = <double>[event.x, event.y, event.z];
           });
         },
         onError: (e) {
@@ -191,10 +191,10 @@ class _MyHomePageState extends State<MyHomePage> {
       ),
     );
     _streamSubscriptions.add(
-      userAccelerometerEvents.listen(
-        (UserAccelerometerEvent event) {
+      gyroscopeEvents.listen(
+        (GyroscopeEvent event) {
           setState(() {
-            _userAccelerometerValues = <double>[event.x, event.y, event.z];
+            _gyroscopeValues = <double>[event.x, event.y, event.z];
           });
         },
         onError: (e) {
